requirements.txt
httpx==0.27.*
beautifulsoup4==4.12.*
lxml==5.*
google-genai>=1.0.0
aiofiles==23.*
aiosqlite==0.20.*
tqdm==4.*
python-dotenv==1.0.*
